+++ src/LL/LLSpaceJapanese.ts
@@ -6,6 +6,6 @@ import ranges2diagnostics from '../util/ranges2diagnostics';
 
 export default function LLSpaceJapanese(doc: vscode.TextDocument, txt: LLText): vscode.Diagnostic[] {
     const code = "LLSpaceJapanese";
-    const ranges = detectMathSpace(doc, txt, "ja");
+    const ranges = detectMathSpace(doc, txt);
     return ranges2diagnostics(code, messages[code], ranges);
 }